fix: tests are moved to tests_csv_reader
